Further scaffolding of our documentation.
We now have all the major parts of the documentation setup up and running.

This commit and a number of READMEs that references each other and outlines the strategy for documentation:
We document the various components of the platform idenpendently, and keep the overall architecture documentation
in a seperate documentation directory. All diagrams are stored in a shared digrams folder and we have CI automation
that renderes the diagrams. We currently support PlantUML and Draw.io aka diagrams.net diagrams.

DDFDPDEL-60