Re-organize our infrastructure as code setup
This commit moves the pieces of our infrastructure setup around so that
it is a bit easier to navigate. One of the primary benefits of the
new layout is that we can place a .dplsh.profile file high enough in the
directory-structure that we can launch a single shell that has access
to all infrastructure tooling and not just terraform.