Add a tool for deploying core test sites
This commit adds the scaffolding for a new set of "dpladm" tools that we
can use to administer sites on hosted on the platform.

It currently includes a single functionality for deploying to a site.
We currently only support deploying to "core test" sites, that is a
simpler version of an editor-site ment to be used to test the
functionality of a dpl-cms release.

DDFDPDEL-102