Start documenting the lifecycle for a site on the platform
This commit adds documentation and tooling for setting up a site.
This is only the start of the work of describing the full lifecycle. We
currently only describe how to set up the site, and have a lot of manuel
steps that may be automated in the future.

The documentation has been written while setting up a site ment for
core tests, and thus does not address tearing down the site, and/or how
to set up backups and other steps one would expect for a production site.

DDFDPDEL-101