Document the full lifecycle of a core test site.
The current process is quite manual and to the point without more considered
steps such as backing up the site before deleting it. This is fine for
core-test sites that are not expected to need a restore
after deletion.

We will iterate on these steps as we get further along with the different
profile-types, and automate the process more and more.

DDFDPDEL-101