Configure Loki, Promtail and Grafana and document their use
These changes follows in the heels of our other support components.
We install them manually via task, but the values-files are templated
and we have some basic options for eg. diffing before applying the chart.

A new storage account is created, and Lokis logs are placed in it.
Prometheus storage is still handled in-cluster, and its, Grafana, and
Lokis block-storage disks are all created bye AKS.

I have added some slightly more robust handling of the installation of
charts, and will be refactoring the rest of the support-scripts in a
future PR.

DDFDPDEL-74