Provision storage for Backup and Lagoon files
This commit introduces two blob containers for lagoon-files and
k8up backups respectively. As we expect to put a MinIO based
proxy in front of the containers to provide S3 compatabillity
we also generate a set of credentials for each container
which will be used by the client while we keep the azure-specific
keys secret.