Ensure we can see the remote IP in the Ingress-Nginx logs
Due to how the load-balancer is set up pr default in kube, we do not
get the clients real ip out of the box. This commit reconfigures the
load-balancer to make the remote ip visible, and adjust the logs to use
another field to log the ip.

See https://docs.microsoft.com/en-us/azure/aks/load-balancer-standard#maintain-the-clients-ip-on-inbound-connections
for some more details on the issue