Document how to download and restore backups.
We wish to have a guide for a stable way of fetching and restoring
backups. This commit adds a runbook that describes this procedure.

As we introduce runbooks that describes one procedure, it becomes clear
that they has to reference other more general procedures. One such case
in this situation was how to run dplsh. So, I've also added a runbook
that describes the "how" of using dplsh. As a part of that effort I've
removed some sections from the infrastructure readme in favor of
referencing the runbook.

DDFDPDEL-116