Change the deployment strategy towards being more declarative
Before this commit, the language around getting a new release deployed
on a site was very imperative. We where doing "a deployment". While this
is still the goal we want to achieve, we also want to stay as declarative
as possible to make it easier to keep a clear record of what we are
changing, and enabling automation as much as possible.

To that effect, this commit

1. Adds a ADR that describes the reasoning behind going with a
   declarative approach
2. Refactors and updates the dpladm "deployer" to now take its
   configuration from sites.yaml and thus becoming more declarative.

These changes are made in combination with the implementation of the
lifecycle for a editor site, so the changes to dpladm does also include
some features such as the support for configuring a sites domain and supporting the use of ssh from within dplsh.

See the subsequent commit for the rest of this functionality

DDFDPDEL-176