Describe the full lifecycle for a site
This commit adds documentation on how to add, deploy to, and remove a
site on the dpl-platform. We already had some documentation for "core-
profile" sites, but we have now reached the point where we can generalize
the administration of sites regardless of what they are used for. In
effect a "core-profile" site is just a site with an autogenerated domain.

A number of minor fixes to the existing documentation was made during the
test of deployment as well.

DDFDPDEL-176