Document how to set environment variables
We already had some documentation on how to set an variable on a lagoon
project in order for the build-process to access private image-registries

This PR builds on that work and generalises the task so that it can
be used to set any environment variable. The task automates the steps
to set an environment variable as documented in https://docs.lagoon.sh/lagoon/using-lagoon-advanced/environment-variables

This commit also simplifies the process a bit by now picking up an users
lagoon bearer token directly from the lagoon login context instead of
having the user perform that step manually.

Last but not least, I add a missing step to the site deploy runbook.