This commit changes two things about how we provision Harbor
We originally installed Harbor as part of installing Lagoon core. A
while back, we realised that Harbor is as seperate from Laggon as eg.
the Ingress Controller, so we might just as well install Harbor along
with the other support projects before we start installing Lagoon.

To get a step closer to this point, the Harbor installation task was
renamed and the documentation updated. But, we still installed the
Harbor directly from the Taskfile and not via a seperate shell-script
like we did with the rest of the support projects.

The first change implemented by this commit is a refactor so that we now
install Harbor via a shell-script using the same pattern as the rest
of the support tools.

The second change is a switch from having Harbors registry store its
images in an in-cluster Persistent Volume, to an externally hosted
Azure Blob Container. To achieve this we expand the dpl-platform-environment
Terraform module to provision the required storage, and then re-configure
Harbor to use it.

DDFDPDEL-196