Increase storage provisioned for mariadb for increased performance
The MariaDB server provisioned to support Lagoon is set up with a
configurable amount of storage. The amount of IOPS the server can perform
against the storage scales linearly with the amount of provisioned
storage.

Before this commit the dpl-platform-environment Terraform module would
default to provision 5GB of storage, which gave us 100IOPS.

This is not enough to support a Drupal site install where the Redis
cache is disabled and the database is used as a fallback cache store.

This commit increases the default to 102400, and updates the current
value for dplplat01 to 409600 which is the result of some experimentation
that showed that 102400 is probably enough going foreward

DDFDPDEL-196