Upgrade to latest version of Loki
The format of the values-files has changed a bit.

We now run with a seperate "read" and "write" pod - this required a duplication of the replica and persistence config.

The path where the volume gets mounted as changed as well, so that required a change to the shipper path.

As we're now switched to a distributed mode, the retention works differently. To accommedate this we now have a root-level
compactor / limits config

Bumped the size of the disk to 16Gi as that what azure is going to give us when we ask for 10Gi