Implement automation for extracting the current and latest version of support workloads
We want to keep track of how far behind we are on keeping things up to date. This can be a somewhat manual process
so to lessen the toil this commit implements an automated process for going through the things we have currently
deployed and getting the current and latest version of the projects.

We're currenly only doing this for the support workloads, and not for Lagoon and AKS.

DDFDRIFT-26