Install openssh in dplsh to ensure we can run ssh-related commands inside it
The guide in [the guide for provisioning Github repos](/docs/runbooks/add-library-site-to-platform.md) step 2 suggests running ssh-agent and ssh-add inside dplsh, but these do not work as openssh is missing from the shell.

To make it easier to run everything inside the shell, we add the missing dependency, so the guide is now correct.