First step towards driving all infra state from sites.yaml
The new task `sites:sync` syncs *all* sites from sites.yaml, ensuring that everything is set up for them as it should.
This includes setting up Lagoon with GITHUB token, capturing the Lagoon deploy key, saving it in sites.yaml, and ensuring all the Github repos are provisioned with Terraform.

A single case is missing right now, which is running a first `lagoon deploy` for new sites. We need to figure out if that is actually required, and if so, if we can detect whether it needs doing