Code deduplication: extract "require_site" precondition
This is an example of how we might approach less duplicated code when trying to ensure variables are correctly defined before executing tasks.

We considered using `requires` (https://taskfile.dev/usage/#ensuring-required-variables-are-set) but it only checks if variables exist, and does not validation on the values, e.g. that they are non-empty. Using that would seem like a regression in our case.

Instead we might consider this approach, where we extract and reuse common preconditions, to ensure that we always validate the same field in the same way? Thoughts welcome