Scale up ingresses more! perhaps
We scaled to max 10 ingresses and the autoscaler immediately went to the limit. We will see if they stay there after the 5 minute stabilization window. If they do, or even if they stay close to (say 8+) it may be worth scaling the resources instead.

This commit suggests how to do that, while also increasing the autoscaling range to a more realistic one for our load.