Scale up ingress-nginx, based on metrics
Based on data from Grafana this seems like a reasonable level to request to be able to deal with spikes in traffic.

The ingresses (15!) are currently using around 300-400 Mi of memory and
barely perceptible CPU. We would like to run on slightly fewer ingresses
with room to scale up, so this request level (1.1Gi) should accomodate
that wish.

We have 16 Gb memory available on each node, and allowing an ingress to
be a memory heavy instance makes sense. Even at this level, we could
accomodate 14 ingresses on a node, which is a lot! So it seems
reasonable.