Task for checking for CAA records in all sites in sites.yaml
Given primary- and secondary-domains in sites.yaml this task checks through all domains to see if any have a CAA record attached. If they do, this is reported.

CAA records present in a DNS record means that we *must* specify the certificate authority allowed to provision certificates for the domain. If any of the libraries have this record we must get them to adjust it prior to going live.