Add task to show the latest backup for each site
Currently we have no good way of identifying failing backups. Our
best way to check that they are running for all sites is to check when
the last succesfull backup was run.

To help in this process we add a new task which retrieves this from
each site in Lagoon and prints it.

Use silent to avoid printing the shell command for each site. This
makes the output more compact.