Cleanup development environments in our core Lagoon project.
With our setup Lagoon creates a development environment automatically
each time a pull request is created. Such an environments should also
be deleted when the pull request is closed. This is usually the case
when the PR is merged but often failed when the pull request is closed
without merging.

To work around this issue this script tries to close all development
environments in Lagoon where the corresponding pull request has been
either closed or merged.