Do not call whoami when configuring Lagoon CLI
Currently this will fail with the error message:
graphql: Cannot read properties of undefined (reading 'realmRoles')

This error is caused by role and membership work attempted in the
Lagoon Keycloak instance.

Calling lagoon whoami is not important in regard to the workings of
the dplsh so we remove this command until the problem has been
resolved.