Add task for restoring database backup to site/environment in Lagoon
This is an initial attempt at scripting the restore process described
in docs/runbooks/retrieve-restore-backup.md.

This is useful in situations where we need to restore a backup into a
site but also if we want to restore backups across projects or
environments e.g. if we need to restore data from a production site
into a test environment. This can be achieved by specifying different
source and target project/environment.

The task has a couple of notable differences in the restore proces:

1. The restore script is not multiline. This hampers readability

2. There is no explicit call to "drush cache:rebuild". Kubernetes
currently OOM kills such attempts for test environments. This is
replaced by a deployment. This process will in practice rebuild the
cache and ensure that any other processed needed to update the
environment are run.

The current implementation works but has room for future expansion:

1. The task currently only deals with database restore. Over time it
could be expanded with a site:backup:restore:files task and then
combined into a site:backup:restore.

2. The task always retrieves the latest backup. This could be expanded
to support a specific backup time.