Do not set task directory when ensuring project
On OSX we see the following error when running deployments:
No Taskfile found at "/home/dplsh/host_mount/environments/dplplat01"

There is no Taskfile in the provided directory and it is not
intentional that we execute tasks from this directory either.

For unknown reasons we currently set the default directory to .dir_env
which is /home/dplsh/host_mount/environments/dplplat01. This should
not be necessary.

The only command which references a specific part of the file system
is the SITE_PLAN var. This already uses .dir_env.

Consequently we remove it.