Upgrade bnf site to use the webmaster plan
We need a site to act as a server when testing the the content
sharing system on staging which in this context acts as a client. To
test recent changes it is usually required that the client and server
both use the newest version.

To provide this we update the bnf site to use the webmaster plan.
We can use the module-test site to act as a server and our current
workflows ensures that it is kept up to date with recent content.

The staging site currently acts as a client.