Keep autogenerated routes for Staging, CMS School and Canary sites
A primary domain has previously been set for these domains to make
cross site authentication work between the CMS and Go sites.

This causes problems as the OpenID clients for Adgangsplatformen used
on these sites are configured to only allow the autogenerated domains
(typically with the varnish service).

Thus we need to keep the autogenerated routes available for
Adgangsplatformen login to work while we wait for these to be updated.