Do not make the BNF site follow the Go pilot schedule by default
This prevents us from from upgrading the moduletest version of this
site individually from all other sites.

This is needed as the BNF moduletest site acts as the server with
the staging site acting as client and we need both client and server
to have the latest versions to test new features.

Apart from changing up anchors this commit should be a no-op.