Add support for defining PHP version for sites
Currently PHP versions are split between multiple repositories:

- Multiple locations in the CMS project
- Here in the corresponding templates

However we want to be able to control which PHP version to use *and*
be able to differentiate on a site and environment level. Default
sites only have a main environment which gets the newest version
of the CMS using the corresponding version of PHP.

Webmaster sites need different versions of PHP for main and moduletest
environments as the moduletest environment may get a CMS version which
also requires a newer version of PHP.

To address this we:

1. Add support for templating the PHP version to use in Docker files
2. Add support for varying this per environment/branch

The actual change is primarily a matter of extracting the PHP version
from sites.yaml and passing it through the stack.

The getPhpVersion() function is added to support retrieving PHP
versions for specific environments e.g. moduletest-php-version or the
fallback php-version valules.

This is then passed along by adding additional variables to function
arguments. The position of the new argument is based on whether the
variable is required or optional.

sites.yaml is updated with the current PHP version, 8.1, to show that
it works.

After this change DIFF=1 SITE=canary task site:sync should not return
a diff.

If you change the main or moduletest PHP version for the site to e.g.
8.3 should be reflected appropriately in the diff.