Don't exclude js/css files and anchor styles exclude
`--exclude=js/*` would also exclude files in modules in modules_local,
causing the modules to fail. Remove the exclude as there doesn't seem
to be any css/js at the root of files anymore?

Also anchor the styles exclude with `/` to ensure it's only a styles
directory in the root that's excluded.